Crate kaleidoscope_macro

Sub-crate for macros I will be using in this repo.

Macros

function_name

Get the name of the function (no paths) of the current function/method.

function_path

Get the full qualified name of the current function/method.

hash_map

Create a hash map

impl_display

Implement a default version of std::fmt::Display for a type if that type already implements std::fmt::Debug.

iterator_to_str

Convert an iterator into a string with each element being a separated by a string.

ok_none

Unwrap an Option but return Ok(None) if the Option is None.

return_ok_some

If $option is Some, return Ok($option).

untrimmed_function_path

See function_path.